third_party/WebKit/LayoutTests/inspector/sources/debugger/async-callstack-scopes-expected.txt


Test that sections representing scopes are expandable and contain correct data for async call frames.

Set timer for test function.
Dumping call frames with scope chains:

1) innerTimeout2 async-callstack-scopes.html:51
Local
    localInInnerTimeout2: "innerTimeout2"
    this: Window
WindowGlobal
    <section collapsed>

2) innerFunction async-callstack-scopes.html:18
ObjectWith Block
    foo: "bar2"
Catch
    e: Error: An exception
Local
    arguments: Arguments[1]
    innerFunctionLocalVar: 202
    this: Window
    x: 200
Closure
    callback: function innerTimeout2()
    makeClosureLocalVar: "local.2"
    n: 2
    withScope: Object
WindowGlobal
    <section collapsed>

3) timeout2 async-callstack-scopes.html:53
Local
    innerTimeout2: function innerTimeout2()
    localInTimeout2: "timeout2"
    this: Window
WindowGlobal
    <section collapsed>

4) innerTimeout1 async-callstack-scopes.html:40
Local
    localInInnerTimeout1: "innerTimeout1"
    this: Window
Closure
    localInTimeout1: "timeout1"
WindowGlobal
    <section collapsed>

5) innerFunction async-callstack-scopes.html:18
Local
    arguments: Arguments[1]
    innerFunctionLocalVar: 102
    this: Window
    x: 100
Closure
    callback: function innerTimeout1()
    makeClosureLocalVar: "local.1"
    n: 1
    withScope: Object
WindowGlobal
    <section collapsed>

6) timeout1 async-callstack-scopes.html:42
Local
    innerTimeout1: function innerTimeout1()
    localInTimeout1: "timeout1"
    this: Window
WindowGlobal
    <section collapsed>

7) testFunction async-callstack-scopes.html:31
Local
    delay1: 0
    this: Window
WindowGlobal
    <section collapsed>
